+++ governance-app/src/main.c
@@ -30,7 +30,6 @@
 #include "signHigherLevelKeyUpdate.h"
 #include "signUpdateAuthorizations.h"
 #include "signUpdateBakerStakeThreshold.h"
-#include "signUpdateElectionDifficulty.h"
 #include "signUpdateExchangeRate.h"
 #include "signUpdateFoundationAccount.h"
 #include "signUpdateGasRewards.h"
@@ -60,7 +59,6 @@ instructionContext global;
 #define INS_UPDATE_GAS_REWARDS           0x23
 #define INS_UPDATE_FOUNDATION_ACCOUNT    0x24
 #define INS_UPDATE_MINT_DISTRIBUTION     0x25
-#define INS_UPDATE_ELECTION_DIFFICULTY   0x26
 #define INS_UPDATE_BAKER_STAKE_THRESHOLD 0x27
 
 #define INS_UPDATE_ROOT_KEYS          0x28
@@ -110,9 +108,6 @@ void handler(
                         break;
                     case INS_UPDATE_MINT_DISTRIBUTION:
                         handleSignUpdateMintDistribution(cdata, p2, flags);
-                        break;
-                    case INS_UPDATE_ELECTION_DIFFICULTY:
-                        handleSignUpdateElectionDifficulty(cdata, flags);
                         break;
                     case INS_UPDATE_BAKER_STAKE_THRESHOLD:
                         handleSignUpdateBakerStakeThreshold(cdata, flags);