+++ governance-app/src/signUpdateGasRewards.c
@@ -13,13 +13,9 @@ UX_STEP_NOCB(
 UX_STEP_NOCB(
     ux_sign_gas_rewards_2_step,
     bnnn_paging,
-    {.title = "Finalization proof", .text = (char *) global.signUpdateGasRewardsContext.gasFinalization});
-UX_STEP_NOCB(
-    ux_sign_gas_rewards_3_step,
-    bnnn_paging,
     {.title = "Account creation", .text = (char *) global.signUpdateGasRewardsContext.gasAccountCreation});
 UX_STEP_NOCB(
-    ux_sign_gas_rewards_4_step,
+    ux_sign_gas_rewards_3_step,
     bnnn_paging,
     {.title = "Chain update", .text = (char *) global.signUpdateGasRewardsContext.gasChainUpdate});
 UX_FLOW(
@@ -28,38 +24,23 @@ UX_FLOW(
     &ux_sign_gas_rewards_1_step,
     &ux_sign_gas_rewards_2_step,
     &ux_sign_gas_rewards_3_step,
-    &ux_sign_gas_rewards_4_step,
     &ux_sign_flow_shared_sign,
     &ux_sign_flow_shared_decline);
 
 void handleSignUpdateGasRewards(uint8_t *cdata, volatile unsigned int *flags) {
     cdata += parseKeyDerivationPath(cdata);
-
     cx_sha256_init(&tx_state->hash);
-    cdata += hashUpdateHeaderAndType(cdata, UPDATE_TYPE_GAS_REWARDS);
+    cdata += hashUpdateHeaderAndType(cdata, UPDATE_TYPE_GAS_REWARDS_CPV2);
 
-    // Baker GAS bytes
     uint32_t gasBaker = U4BE(cdata, 0);
-    fractionToPercentageDisplay(ctx->gasBaker, sizeof(ctx->gasBaker), gasBaker);
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
-    cdata += 4;
+    uint32_t gasAccountCreation = U4BE(cdata, 4);
+    uint32_t gasChainUpdate = U4BE(cdata, 8);
 
-    // Finalization proof GAS bytes
-    uint32_t gasFinalizationProof = U4BE(cdata, 0);
-    fractionToPercentageDisplay(ctx->gasFinalization, sizeof(ctx->gasFinalization), gasFinalizationProof);
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
-    cdata += 4;
+    updateHash((cx_hash_t *) &tx_state->hash, cdata, 12);
 
-    // Account creation GAS bytes
-    uint32_t gasAccountCreation = U4BE(cdata, 0);
+    fractionToPercentageDisplay(ctx->gasBaker, sizeof(ctx->gasBaker), gasBaker);    
     fractionToPercentageDisplay(ctx->gasAccountCreation, sizeof(ctx->gasAccountCreation), gasAccountCreation);
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
-    cdata += 4;
-
-    // Chain update GAS bytes
-    uint32_t gasChainUpdate = U4BE(cdata, 0);
     fractionToPercentageDisplay(ctx->gasChainUpdate, sizeof(ctx->gasChainUpdate), gasChainUpdate);
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
 
     ux_flow_init(0, ux_sign_gas_rewards, NULL);
     *flags |= IO_ASYNCH_REPLY;