+++ governance-app/src/signUpdateGasRewards.h
@@ -10,7 +10,6 @@ void handleSignUpdateGasRewards(uint8_t *cdata, volatile unsigned int *flags);
 
 typedef struct {
     uint8_t gasBaker[8];
-    uint8_t gasFinalization[8];
     uint8_t gasAccountCreation[8];
     uint8_t gasChainUpdate[8];
 } signUpdateGasRewardsContext_t;